+++ condow_core/src/download_session.rs
@@ -5,7 +5,7 @@ use futures::future::BoxFuture;
 
 use crate::{
     condow_client::CondowClient,
-    errors::{CondowError, GetSizeError},
+    errors::CondowError,
     reporter::{CompositeReporter, NoReporting, Reporter, ReporterFactory},
     streams::{ChunkStream, PartStream},
     Condow, DownloadRange, Downloads, GetSizeMode, StreamWithReport,
@@ -174,7 +174,7 @@ impl<C: CondowClient, RF: ReporterFactory> DownloadSession<C, RF> {
     }
 
     /// Get the size of a file at the BLOB at location
-    pub async fn get_size(&self, location: C::Location) -> Result<usize, GetSizeError> {
+    pub async fn get_size(&self, location: C::Location) -> Result<usize, CondowError> {
         self.condow.get_size(location).await
     }
 }