+++ condow_core/src/errors.rs
@@ -1,46 +1,76 @@
 //! Error types returned by Condow
+use std::fmt;
+
 use thiserror::Error;
 
 #[derive(Error, Debug)]
-pub enum CondowError {
-    #[error("invalid range: {0}")]
-    InvalidRange(String),
-    #[error("not found: {0}")]
-    NotFound(String),
-    #[error("access denied: {0}")]
-    AccessDenied(String),
-    #[error("error: {0}")]
-    Remote(String),
-    #[error("io error: {0}")]
-    Io(String),
-    #[error("error: {0}")]
-    Other(String),
+pub struct CondowError {
+    msg: String,
+    #[source]
+    source: Option<anyhow::Error>,
+    kind: CondowErrorKind,
 }
 
-impl From<GetSizeError> for CondowError {
-    fn from(dfe: GetSizeError) -> Self {
-        match dfe {
-            GetSizeError::NotFound(msg) => CondowError::NotFound(msg),
-            GetSizeError::AccessDenied(msg) => CondowError::AccessDenied(msg),
-            GetSizeError::Remote(msg) => CondowError::Remote(msg),
-            GetSizeError::Io(msg) => CondowError::Io(msg),
-            GetSizeError::Other(msg) => CondowError::Other(msg),
+impl CondowError {
+    pub fn new<T: Into<String>>(msg: T, kind: CondowErrorKind) -> Self {
+        Self {
+            msg: msg.into(),
+            source: None,
+            kind,
         }
+    }
+    pub fn new_invalid_range<T: Into<String>>(msg: T) -> Self {
+        Self::new(msg, CondowErrorKind::InvalidRange)
+    }
+
+    pub fn new_not_found<T: Into<String>>(msg: T) -> Self {
+        Self::new(msg, CondowErrorKind::NotFound)
+    }
+
+    pub fn new_access_denied<T: Into<String>>(msg: T) -> Self {
+        Self::new(msg, CondowErrorKind::AccessDenied)
+    }
+
+    pub fn new_remote<T: Into<String>>(msg: T) -> Self {
+        Self::new(msg, CondowErrorKind::Remote)
+    }
+
+    pub fn new_io<T: Into<String>>(msg: T) -> Self {
+        Self::new(msg, CondowErrorKind::Io)
+    }
+
+    pub fn new_other<T: Into<String>>(msg: T) -> Self {
+        Self::new(msg, CondowErrorKind::Other)
+    }
+
+    pub fn with_source<E: Into<anyhow::Error>>(mut self, err: E) -> Self {
+        self.source = Some(err.into());
+        self
+    }
+
+    pub fn msg(&self) -> &str {
+        &self.msg
+    }
+
+    pub fn kind(&self) -> CondowErrorKind {
+        self.kind
     }
 }
 
-#[derive(Error, Debug)]
-pub enum GetSizeError {
-    #[error("not found: {0}")]
-    NotFound(String),
-    #[error("access denied: {0}")]
-    AccessDenied(String),
-    #[error("error: {0}")]
-    Remote(String),
-    #[error("io error: {0}")]
-    Io(String),
-    #[error("error: {0}")]
-    Other(String),
+impl fmt::Display for CondowError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", self.msg)
+    }
+}
+
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub enum CondowErrorKind {
+    InvalidRange,
+    NotFound,
+    AccessDenied,
+    Remote,
+    Io,
+    Other,
 }
 #[derive(Error, Debug)]
 #[error("io error: {0}")]