+++ condow_core/CHANGELOG.md
@@ -10,6 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * `CompositeReporter`
 
+### CHANGED
+
+* `CondowError`
+
+### Removed
+
+* `GetSizeError`
+
 ## [0.5.0] - 2021/08/21
 
 ### ADDED